logistic-regression/requirements.txt
Flask==2.2.2
numpy==1.24.1
pillow==9.4.0